% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manager.R
\name{DataManager}
\alias{DataManager}
\title{DataManager R6 Class}
\description{
An R6 class for managing and processing energy data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data frame containing the processed energy data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DataManager-new}{\code{DataManager$new()}}
\item \href{#method-DataManager-filter_data}{\code{DataManager$filter_data()}}
\item \href{#method-DataManager-calculate_kpis}{\code{DataManager$calculate_kpis()}}
\item \href{#method-DataManager-clone}{\code{DataManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-new}{}}}
\subsection{Method \code{new()}}{
Initializes the DataManager object with a data frame.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$new(df)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A data frame to be processed. It should contain columns
`date`, `site`, `type`, `value`, and `carbon emission in kgco2e`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-filter_data"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-filter_data}{}}}
\subsection{Method \code{filter_data()}}{
Filters the data based on a selected date range and sites.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$filter_data(start_date, end_date, selected_sites)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The start date of the filter range.}

\item{\code{end_date}}{The end date of the filter range.}

\item{\code{selected_sites}}{A vector of site names to filter by.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A filtered data frame.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-calculate_kpis"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-calculate_kpis}{}}}
\subsection{Method \code{calculate_kpis()}}{
Calculates Key Performance Indicators (KPIs) from the data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$calculate_kpis()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing the calculated KPIs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DataManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DataManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
