% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SAMPLE_ASSIGNMENT_DATA}
\alias{SAMPLE_ASSIGNMENT_DATA}
\title{Sample Assignment Data}
\format{
A data frame with 2973 rows and 6 variables:
\describe{
  \item{id}{Unique identifier for each entry.}
  \item{site}{The site of the energy consumption.}
  \item{date}{The date of the data entry.}
  \item{type}{The type of energy consumption (e.g., Water, Electricity).}
  \item{value}{The energy usage value.}
  \item{carbon emission in kgco2e}{The carbon emission value in kg.}
}
}
\usage{
SAMPLE_ASSIGNMENT_DATA
}
\description{
A dataset containing sample energy consumption and carbon emission data.
This data is used for the shiny app dashboard.
}
\keyword{datasets}
